Refactor home page, update typography, and simplify font imports
